requirements.txt
pyrogram==1.4.16
tgcrypto==1.2.5
python-dotenv==1.0.0
aiofiles==23.2.1
